use_index now requires libmswstr. See #215 and #226
Print a warning if use_index is turned on but libmswstr is not found.  I
suppose we could enable indexes only on JET3 databases, but I am not at
all confident in the JET3 index logic, as it seems to break on non-ASCII
input. So I'd rather just print this warning and require some
hoop-jumping.