Move includes out of extern "C" block (#435)
"mbdtools.h" couldn't be included from C++ files, because the "glib.h"
include is in the extern "C" block. The latter includes C++ headers
for C++ source files, which define templates. These don't work with
extern "C" linkage and the compilation falls apart with numerous
errors.

To fix this problem, move the includes out of the extern "C" block.